FIX: Python library linking - not using 3.13.4 (#74)
### Summary  
This pull request updates the Python version used in the build and
validation pipelines to address a known issue with Python 3.13.4 and
ensure compatibility with Python 3.13.5. The changes include adjustments
to the version specifications and relevant comments for clarity.

Pipeline updates:

*
[`eng/pipelines/build-whl-pipeline.yml`](diffhunk://#diff-a871d64acfb78366b4b077045db53551c2df7d622854569f5f6780ad2ec6f911L53-R58):
Updated the `pythonVersion` for `py313_x64` and `py313_arm64` jobs from
`3.13` to `3.13.5` to ensure the use of the latest compatible version.
*
[`eng/pipelines/pr-validation-pipeline.yml`](diffhunk://#diff-296c8f902bbd70f34ee1c8c32383c8c99165fe4c8e5b0f234f8f22246e56a621L17-R21):
Updated the `versionSpec` in the `UsePythonVersion` task to `3.13.5` and
added a comment explaining the temporary workaround due to ADO indexing
delays and known issues with Python 3.13.4.

### Issue Reference  
Fixes
[AB#37710](https://sqlclientdrivers.visualstudio.com/c6d89619-62de-46a0-8b46-70b92a84d85e/_workitems/edit/37710)

### Checklist  
- [x] **Tests Passed** (if applicable)  
- [x] **Code is formatted**   
- [x] **Docs Updated** (if necessary) 

### Testing Performed  
<!-- How was this fix tested? -->
- [x] Unit Tests
- [x] Manual Testing  
  - [x] Python Version: `3.13.3`
  - [x] OS: `Windows`