FIX: validate numeric data for range (upper and lower bound) (#77)
[AB#37475](https://sqlclientdrivers.visualstudio.com/c6d89619-62de-46a0-8b46-70b92a84d85e/_workitems/edit/37475)
This pull request enhances the `BindParameters` function in
`mssql_python/pybind/ddbc_bindings.cpp` by introducing range validation
for various parameter types to ensure data integrity and prevent
out-of-range errors. Below are the key changes grouped by parameter
type:

### Integer Parameter Validations

* **Signed 16-bit integer (`short`)**: Added range validation to ensure
the value is within the limits of a signed 16-bit integer. Throws an
exception if the value is out of range.
* **Unsigned 16-bit integer (`unsigned short`)**: Added validation to
check that the value does not exceed the maximum limit of an unsigned
16-bit integer.
* **Signed 64-bit integer (`int64_t`)**: Introduced range validation to
confirm the value is within the bounds of a signed 64-bit integer.
* **Unsigned 64-bit integer (`uint64_t`)**: Added a check to ensure the
value does not exceed the maximum limit for an unsigned 64-bit integer.

### Date Parameter Validation

* **SQL Server Date Range**: Implemented validation for `year`
attributes in date parameters to ensure they fall within the valid SQL
Server date range (1753-9999).

### Checklist  
- [x] **Tests Passed** (if applicable)  : All pytests passed.