RELEASE: 0.6.0 (#89)
### Summary  
This pull request introduces updates to the `mssql-python` project,
including versioning changes, platform support enhancements, and
pipeline adjustments. The most important changes involve updating the
version to `0.6.0`, clarifying macOS support, and commenting out a
release task in the pipeline for testing purposes.

### Versioning and Feature Updates:
*
[`setup.py`](diffhunk://#diff-60f61ab7a8d1910d86d9fda2261620314edcae5894d5aaa236b821c7256badd7L89-R89):
Updated the version of the `mssql-python` package from `0.5.0` to
`0.6.0` to reflect the latest changes and improvements.

*
[`PyPI_Description.md`](diffhunk://#diff-5236254592b2fea0773f17424b16acf82e3aa351ad8bae33871de5c98eb76eedL12-R13):
Enhanced macOS support to include Intel-based systems in addition to
ARM-based systems (M-series). Added clarification that the connection
pooling feature is currently available only on Windows.

### Pipeline Adjustments:
*
[`eng/pipelines/dummy-release-pipeline.yml`](diffhunk://#diff-de3eb3d456dcb54748a65dfec8142a203b22d898a7f95930cfaffc5557d8d134L30-R51):
Commented out the `ESRP Release` task, which was previously configured
for a dummy Maven release for testing purposes. This change prevents the
task from being executed while retaining the configuration for future
reference.