FEAT: Create devskim.yml (#87)
### Summary  

This pull request introduces a new GitHub Actions workflow to integrate
the DevSkim static analysis tool into the CI/CD pipeline. The workflow
is designed to scan code for security issues and upload the results to
the GitHub Security tab.

### New Workflow Integration:

*
[`.github/workflows/devskim.yml`](diffhunk://#diff-cf9c7e0ae50cd2d73f48fc95d5f4784c87b4932f23c9bf60f2223ef12262c4cdR1-R34):
Added a new workflow named "DevSkim" that runs on `push` and
`pull_request` events for the `main` branch, as well as on a weekly
schedule. It uses the `microsoft/DevSkim-Action` to perform security
scans and uploads the results in SARIF format using the
`github/codeql-action/upload-sarif` action.

### Issue Reference  
Fixes:
[AB#37674](https://sqlclientdrivers.visualstudio.com/c6d89619-62de-46a0-8b46-70b92a84d85e/_workitems/edit/37674)

Co-authored-by: Gaurav Sharma <sharmag@microsoft.com>