FIX: Adding row string (#136)
### ADO Work Item Reference  
<!-- Insert your ADO Work Item ID below (e.g. AB#37452) -->
>
[AB#37929](https://sqlclientdrivers.visualstudio.com/c6d89619-62de-46a0-8b46-70b92a84d85e/_workitems/edit/37929)
-------------------------------------------------------------------
### Summary   
This pull request modifies the `Row` class in `mssql_python/row.py` to
improve its string representation methods. The changes include adding a
`__str__` method for user-friendly output and updating the `__repr__`
method to focus on debugging.

Enhancements to string representation:

* Added a `__str__` method to provide a user-friendly string
representation of the `Row` object by returning the tuple of its values
as a string.
* Updated the `__repr__` method to provide a more detailed
representation for debugging purposes by using the `repr` function on
the tuple of values instead of a formatted string.

---------

Co-authored-by: Jahnvi Thakkar <jathakkar@microsoft.com>