Merged PR 6424: Add the baseline and suppression file to guide Guardian
Add the baseline and suppression file to guide Guardian.

The guide is found here https://eng.ms/docs/coreai/devdiv/one-engineering-system-1es/1es-docs/1es-pipeline-templates/features/sdlanalysis/baselines

The baseline and suppression file tell Guardian about the expected results of the scan.
For credential scanning, the signatures of the credentials are coded in the guardian files.

The files were downloaded from the failing build at https://sqlclientdrivers.visualstudio.com/mssql-python/_build/results?buildId=129570&view=results

The process is to run the pipeline which will cause the failure. After doing due diligence to the credentials and making sure that they are safe to have in the source code, the baseline and suppression files can be updated.

Related work items: #39894