FIX: Update  escaping rules in connection string parser and builder (#364)
### Work Item / Issue Reference  

<!-- External contributors: GitHub Issue -->
> GitHub Issue: #363

-------------------------------------------------------------------
### Summary   
<!-- Insert your summary of changes below. Minimum 10 characters
required. -->

FIX: The escaping rules for the parser and the builder

The parser was treating {{ as an escaped string and was parsing it as {
and then the builder was escaping { to {{
However the ODBC driver doesn't require escaping of the opening brace,
if its already wrapped in {}. Only closing brace values inside curlies
require escaping. Hence a literal pw}d needs to be sent as {pw}}d}

Fixed the parser and the builder to take care of this nuance.

Fixes #363