FIX: Invalid data type for None (#332)
<!-- mssql-python maintainers: ADO Work Item -->
>
[AB#39794](https://sqlclientdrivers.visualstudio.com/c6d89619-62de-46a0-8b46-70b92a84d85e/_workitems/edit/39794)

<!-- External contributors: GitHub Issue -->
> GitHub Issue: #273

-------------------------------------------------------------------
### Summary   
<!-- Insert your summary of changes below. Minimum 10 characters
required. -->
This pull request adds support for binding arrays of NULL parameters in
SQL statements and introduces a corresponding test to ensure correct
behavior when executing multiple inserts with only NULL values. The main
focus is on handling the `SQL_C_DEFAULT` type for parameter arrays,
ensuring all values are NULL, and verifying this functionality through
unit testing.

**Parameter binding improvements:**

* Added logic to `BindParameterArray` in `ddbc_bindings.cpp` to handle
the `SQL_C_DEFAULT` type, ensuring that arrays of NULL values are
properly bound and validated. If any non-NULL value is detected, an
exception is thrown.

**Testing enhancements:**

* Added a new test `test_executemany_NONE_parameter_list` in
`test_004_cursor.py` to verify that `executemany` correctly inserts rows
with all NULL values and that the count matches the expected number of
inserts.